Modify _print.c

+++ p_string.c
@@ -3,10 +3,10 @@
 /**
  * p_string - print a string to the stdio
  * @args: string to print
- * Return - number of bytes printed
+ * Return: number of bytes printed
 */
 
-ssize_t p_string(va_list args)
+int p_string(va_list args)
 {
 	char *str = va_arg(args, char *);
